==> data_cleaning/12_as factor.R <==
# set wd
setwd("/rds/general/user/sv619/projects/hda-22-23/live/TDS/Group8_Working/analysis/DATA")

## FEMLAE_TRAIN_DATA_2403

# data MAKE CATEGORICAL VARIABLES AS FACTS FEMALE
female_train_data_1103$place_of_birth_N <- as.factor(female_train_data_1103$place_of_birth_N)
female_train_data_1103$place_of_birth_E <- as.factor(female_train_data_1103$place_of_birth_E)
female_train_data_1103$accommodation <- as.factor(female_train_data_1103$accommodation)
female_train_data_1103$own_or_rent <- as.factor(female_train_data_1103$own_or_rent)
female_train_data_1103$num_in_house <- as.factor(female_train_data_1103$num_in_house)
female_train_data_1103$avg_house_income <- as.factor(female_train_data_1103$avg_house_income)
female_train_data_1103$sleep_duration <- as.factor(female_train_data_1103$sleep_duration)
female_train_data_1103$smoke_household <- as.factor(female_train_data_1103$smoke_household)
female_train_data_1103$other_medcond <- as.factor(female_train_data_1103$other_medcond)
female_train_data_1103$age_period_start <- as.factor(female_train_data_1103$age_period_start)
female_train_data_1103$menopause <- as.factor(female_train_data_1103$menopause)
female_train_data_1103$number_live_births <- as.factor(female_train_data_1103$number_live_births)
female_train_data_1103$ever_stillbirth_miscarriage_abortion <- as.factor(female_train_data_1103$ever_stillbirth_miscarriage_abortion)
female_train_data_1103$ever_cont <- as.factor(female_train_data_1103$ever_cont)
female_train_data_1103$age_first_cont <- as.factor(female_train_data_1103$age_first_cont)
female_train_data_1103$age_last_cont <- as.factor(female_train_data_1103$age_last_cont)
female_train_data_1103$ever_HRT <- as.factor(female_train_data_1103$ever_HRT)
female_train_data_1103$Education_level <- as.factor(female_train_data_1103$Education_level)
female_train_data_1103$stress_factor <- as.factor(female_train_data_1103$stress_factor)
female_train_data_1103$Family_history_father <- as.factor(female_train_data_1103$Family_history_father)
female_train_data_1103$Family_history_mother <- as.factor(female_train_data_1103$Family_history_mother)
female_train_data_1103$Smoke_status <- as.factor(female_train_data_1103$Smoke_status)
female_train_data_1103$Alcohol_status <- as.factor(female_train_data_1103$Alcohol_status)
female_train_data_1103$urban_rural <- as.factor(female_train_data_1103$urban_rural)
female_train_data_1103$Ever_smoked <- as.factor(female_train_data_1103$Ever_smoked)
female_train_data_1103$ethnicity <- as.factor(female_train_data_1103$ethnicity)
female_train_data_1103$close_major_road <- as.factor(female_train_data_1103$close_major_road)

saveRDS(female_train_data_1103, file = "female_train_1603.rds")

### FEMALE TEST
female_test_data_1103$place_of_birth_N <- as.factor(female_test_data_1103$place_of_birth_N)
female_test_data_1103$place_of_birth_E <- as.factor(female_test_data_1103$place_of_birth_E)
female_test_data_1103$accommodation <- as.factor(female_test_data_1103$accommodation)
female_test_data_1103$own_or_rent <- as.factor(female_test_data_1103$own_or_rent)
female_test_data_1103$num_in_house <- as.factor(female_test_data_1103$num_in_house)
female_test_data_1103$avg_house_income <- as.factor(female_test_data_1103$avg_house_income)
female_test_data_1103$sleep_duration <- as.factor(female_test_data_1103$sleep_duration)
female_test_data_1103$smoke_household <- as.factor(female_test_data_1103$smoke_household)
female_test_data_1103$other_medcond <- as.factor(female_test_data_1103$other_medcond)
female_test_data_1103$age_period_start <- as.factor(female_test_data_1103$age_period_start)
female_test_data_1103$menopause <- as.factor(female_test_data_1103$menopause)
female_test_data_1103$number_live_births <- as.factor(female_test_data_1103$number_live_births)
female_test_data_1103$ever_stillbirth_miscarriage_abortion <- as.factor(female_test_data_1103$ever_stillbirth_miscarriage_abortion)
female_test_data_1103$ever_cont <- as.factor(female_test_data_1103$ever_cont)
female_test_data_1103$age_first_cont <- as.factor(female_test_data_1103$age_first_cont)
female_test_data_1103$age_last_cont <- as.factor(female_test_data_1103$age_last_cont)
female_test_data_1103$ever_HRT <- as.factor(female_test_data_1103$ever_HRT)
female_test_data_1103$Education_level <- as.factor(female_test_data_1103$Education_level)
female_test_data_1103$stress_factor <- as.factor(female_test_data_1103$stress_factor)
female_test_data_1103$Family_history_father <- as.factor(female_test_data_1103$Family_history_father)
female_test_data_1103$Family_history_mother <- as.factor(female_test_data_1103$Family_history_mother)
female_test_data_1103$Smoke_status <- as.factor(female_test_data_1103$Smoke_status)
female_test_data_1103$Alcohol_status <- as.factor(female_test_data_1103$Alcohol_status)
female_test_data_1103$urban_rural <- as.factor(female_test_data_1103$urban_rural)
female_test_data_1103$Ever_smoked <- as.factor(female_test_data_1103$Ever_smoked)
female_test_data_1103$ethnicity <- as.factor(female_test_data_1103$ethnicity)
female_test_data_1103$close_major_road <- as.factor(female_test_data_1103$close_major_road)

saveRDS(female_test_data_1103, file = "female_test_1603.rds")


### MALE TRAIN

male_train_data_1103$place_of_birth_N <- as.factor(male_train_data_1103$place_of_birth_N)
male_train_data_1103$place_of_birth_E <- as.factor(male_train_data_1103$place_of_birth_E)
male_train_data_1103$accommodation <- as.factor(male_train_data_1103$accommodation)
male_train_data_1103$own_or_rent <- as.factor(male_train_data_1103$own_or_rent)
male_train_data_1103$num_in_house <- as.factor(male_train_data_1103$num_in_house)
male_train_data_1103$avg_house_income <- as.factor(male_train_data_1103$avg_house_income)
male_train_data_1103$sleep_duration <- as.factor(male_train_data_1103$sleep_duration)
male_train_data_1103$smoke_household <- as.factor(male_train_data_1103$smoke_household)
male_train_data_1103$other_medcond <- as.factor(male_train_data_1103$other_medcond)
male_train_data_1103$Education_level <- as.factor(male_train_data_1103$Education_level)
male_train_data_1103$stress_factor <- as.factor(male_train_data_1103$stress_factor)
male_train_data_1103$Family_history_father <- as.factor(male_train_data_1103$Family_history_father)
male_train_data_1103$Family_history_mother <- as.factor(male_train_data_1103$Family_history_mother)
male_train_data_1103$Smoke_status <- as.factor(male_train_data_1103$Smoke_status)
male_train_data_1103$Alcohol_status <- as.factor(male_train_data_1103$Alcohol_status)
male_train_data_1103$urban_rural <- as.factor(male_train_data_1103$urban_rural)
male_train_data_1103$Ever_smoked <- as.factor(male_train_data_1103$Ever_smoked)
male_train_data_1103$ethnicity <- as.factor(male_train_data_1103$ethnicity)
male_train_data_1103$close_major_road <- as.factor(male_train_data_1103$close_major_road)

saveRDS(male_train_data_1103, file = "male_train_1603.rds")

### male test
male_test_data_1103$place_of_birth_N <- as.factor(male_test_data_1103$place_of_birth_N)
male_test_data_1103$place_of_birth_E <- as.factor(male_test_data_1103$place_of_birth_E)
male_test_data_1103$accommodation <- as.factor(male_test_data_1103$accommodation)
male_test_data_1103$own_or_rent <- as.factor(male_test_data_1103$own_or_rent)
male_test_data_1103$num_in_house <- as.factor(male_test_data_1103$num_in_house)
male_test_data_1103$avg_house_income <- as.factor(male_test_data_1103$avg_house_income)
male_test_data_1103$sleep_duration <- as.factor(male_test_data_1103$sleep_duration)
male_test_data_1103$smoke_household <- as.factor(male_test_data_1103$smoke_household)
male_test_data_1103$other_medcond <- as.factor(male_test_data_1103$other_medcond)
male_test_data_1103$Education_level <- as.factor(male_test_data_1103$Education_level)
male_test_data_1103$stress_factor <- as.factor(male_test_data_1103$stress_factor)
male_test_data_1103$Family_history_father <- as.factor(male_test_data_1103$Family_history_father)
male_test_data_1103$Family_history_mother <- as.factor(male_test_data_1103$Family_history_mother)
male_test_data_1103$Smoke_status <- as.factor(male_test_data_1103$Smoke_status)
male_test_data_1103$Alcohol_status <- as.factor(male_test_data_1103$Alcohol_status)
male_test_data_1103$urban_rural <- as.factor(male_test_data_1103$urban_rural)
male_test_data_1103$Ever_smoked <- as.factor(male_test_data_1103$Ever_smoked)
male_test_data_1103$ethnicity <- as.factor(male_test_data_1103$ethnicity)
male_test_data_1103$close_major_road <- as.factor(male_test_data_1103$close_major_road)

saveRDS(male_test_data_1103, file = "male_test_1603.rds")
